ForumIngestor watchs mong and index to elastic
